# Conflicts:
#	packages/stripe/pubspec.yaml
#	packages/stripe_android/android/build.gradle
#	packages/stripe_platform_interface/lib/src/models/ach_params.freezed.dart
#	packages/stripe_platform_interface/lib/src/models/ach_params.g.dart

+++ packages/stripe/lib/src/widgets/apple_pay_button.dart
@@ -17,7 +17,7 @@ class ApplePayButton extends StatelessWidget {
     Key? key,
     this.style = PlatformButtonStyle.automatic,
     this.type = PlatformButtonType.plain,
-    this.cornerRadius = 4.0,
+    this.cornerRadius=4,
     this.onPressed,
     double? width,
     double? height = _kApplePayButtonDefaultHeight,
@@ -50,7 +50,7 @@ class ApplePayButton extends StatelessWidget {
   /// Modifies the **corner radius** of the payment button.
   /// To remove the rounded courners, set this value to 0.0.
   /// The default value is set to 4.0
-  final double cornerRadius;
+  final int cornerRadius;
 
   /// Callback that is executed when the button is pressed.
   final VoidCallback? onPressed;
@@ -116,7 +116,7 @@ class _UiKitApplePayButton extends StatefulWidget {
     Key? key,
     required this.style,
     required this.type,
-    this.cornerRadius = 4.0,
+    this.cornerRadius = 4,
     this.onPressed,
     this.onShippingContactSelected,
     this.onCouponCodeEntered,
@@ -126,7 +126,7 @@ class _UiKitApplePayButton extends StatefulWidget {
 
   final PlatformButtonStyle style;
   final PlatformButtonType type;
-  final double cornerRadius;
+  final int cornerRadius;
   final VoidCallback? onPressed;
   final OnDidSetShippingContact? onShippingContactSelected;
   final OnDidSetShippingMethod? onShippingMethodSelected;
